[CLI] Make startup more responsive in the error case, use a volume for postgres, improve logging
