+++ developer-docs-site/scripts/additional_dict.txt
@@ -602,6 +602,7 @@ subtype
 subtyping
 superset
 svg
+symlink
 synchronizer
 terraform
 testnet
